separate gen and def
